chatbot/requirements.txt
langchain
langchain-core
langchain-google-genai
langchain-community
google-generativeai
tavily-python
python-dotenv
fastapi
uvicorn
